proctor-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
onnxruntime>=1.17.0
numpy>=1.26.0
Pillow>=10.2.0
python-multipart>=0.0.9
pydantic>=2.6.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.28
python-dotenv>=1.0.0
httpx>=0.27.0
